This is my re-write into Lit.  Right now I'm stymied:
According to the experts (including Justin Fagnani, the guy who actually wrote Lit),
playing with the InnerHTML should not affect the render process, but it's clear that
whenever I touch the innerHTML field the tiles all re-render because their inner
behavior, their tilt, changes whenever I try to run 'hearts' on the board.

Gonna back-burner this for awhile, because I'm out of ideas and I have other things
to work on while this one percolates in the back-brain.